+++ packages/bridge-controller/src/selectors.ts
@@ -7,6 +7,7 @@ import type {
 import type { GasFeeEstimates } from '@metamask/gas-fee-controller';
 import type { CaipAssetType } from '@metamask/utils';
 import { isStrictHexString } from '@metamask/utils';
+import { BigNumber } from 'bignumber.js';
 import { orderBy } from 'lodash';
 import {
   createSelector as createSelector_,
@@ -24,6 +25,7 @@ import type {
 import { RequestStatus, SortOrder } from './types';
 import {
   getNativeAssetForChainId,
+  isCrossChain,
   isNativeAddress,
   isSolanaChainId,
 } from './utils/bridge';
@@ -404,3 +406,27 @@ export const selectBridgeQuotes = createStructuredBridgeSelector({
   quotesInitialLoadTimeMs: (state) => state.quotesInitialLoadTime,
   isQuoteGoingToRefresh: selectIsQuoteGoingToRefresh,
 });
+
+type MaxBalanceButtonVisibilityClientParams = {
+  isStxEnabled: boolean;
+  balanceValue: string;
+};
+export const selectMaxBalanceButtonVisibilityForSrcToken = createBridgeSelector(
+  [
+    (state) => state.quoteRequest.srcTokenAddress,
+    (state) => state.quoteRequest.srcChainId,
+    (state) => state.quoteRequest.destChainId,
+    (_, { isStxEnabled }: MaxBalanceButtonVisibilityClientParams) =>
+      isStxEnabled,
+    (_, { balanceValue }: MaxBalanceButtonVisibilityClientParams) =>
+      balanceValue,
+  ],
+
+  (srcTokenAddress, srcChainId, destChainId, isStxEnabled, balanceValue) =>
+    Boolean(
+      srcChainId &&
+        !isCrossChain(srcChainId, destChainId) &&
+        (isNativeAddress(srcTokenAddress) ? isStxEnabled : true) &&
+        new BigNumber(balanceValue ?? 0).gt(0),
+    ),
+);